main/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-jose
Jinja2
python-multipart